Shutdown node gracefully from the app before quit
There is no benefit shutting it down from the supervisor (which is more complex and involves making a call).
If the app is forcibly closed, the supervisor will not receive the 'kill' instruction anyway because the shutdown is immediate. Therefore, it makes sense to simply shut down the node directly from the app. The logic of the shutdown is similar to the 'kill' instruction of the supervisor:
- The '/stop' call is issued via API
- If within 3000ms it does not shutdown, it is killed using PID which was stored when spawning the node during the startup phase